extract db support ansi sql92
