+++ db-extractor-core/src/test/resources/ch/adesso/dbextractor/core/DbSupportHsqlDbIT.data.sql
